vcx/requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==5.1.1
djangorestframework==3.15.2
flake8==7.1.1
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.3
sqlparse==0.5.1
